refactor(voc-review): split VocReviewSections into VocAttachmentSection and VocHistoryTab
